Support deduplicating text nodes
In some odd cases with large amounts of data we see cases where the underlying parser starts producing duplicated text nodes (as well as the already handled duplicated elements).
This may be something to do with multiple attachments of events, but for now we can at least just deduplicate the output.